requirements.txt
# Core Dependencies
openai>=1.0.0
whisper-openai>=20231117
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
schedule>=1.2.0

# Audio Processing
pyaudio>=0.2.11
wave>=0.0.1
pydub>=0.25.1

# Email & Communication
smtplib2>=0.2.0
email-validator>=2.0.0
jinja2>=3.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
sqlite3

# Calendar Integration
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
microsoft-graph>=1.0.0

# Web Interface (Optional)
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Task Scheduling
apscheduler>=3.10.0
celery>=5.3.0
redis>=4.6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
tqdm>=4.65.0
loguru>=0.7.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0